Fixed compile error
fixed error with double definiton
